refactor(datatable): split EditorNav from DataTableInner
